Refactor duplicated team sync logic into shared service
