+++ backend/src/main/java/com/predictorama/backend/domain/service/MatchResultSyncService.java
@@ -5,7 +5,6 @@ import com.predictorama.backend.domain.entity.Team;
 import com.predictorama.backend.domain.entity.Tournament;
 import com.predictorama.backend.domain.port.external.FootballDataPort;
 import com.predictorama.backend.domain.port.persistence.MatchRepositoryPort;
-import com.predictorama.backend.domain.port.persistence.TeamRepositoryPort;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,70 +20,15 @@ public class MatchResultSyncService {
     private final PredictionFixtureImportService predictionFixtureImportService;
     private final FootballDataPort footballDataPort;
     private final MatchRepositoryPort matchRepositoryPort;
-    private final TeamRepositoryPort teamRepositoryPort;
+    private final TeamSyncService teamSyncService;
     private final CompetitionCatalog competitionCatalog;
     private final PredictionScoringService predictionScoringService;
 
     private static final Logger log = LoggerFactory.getLogger(MatchResultSyncService.class);
 
-    private Team saveOrGetTeam(Team incomingTeam) {
-        if (incomingTeam == null || isBlank(incomingTeam.getExternalId())) {
-            throw new IllegalArgumentException("Cannot save team with missing externalId");
-        }
-
-        if (isBlank(incomingTeam.getName())) {
-            throw new IllegalArgumentException("Cannot save team with missing name");
-        }
-
-        return teamRepositoryPort.findByExternalId(incomingTeam.getExternalId())
-                .map(existingTeam -> {
-                    boolean teamChanged =
-                            !incomingTeam.getName().equals(existingTeam.getName()) ||
-                            (incomingTeam.getImageUrl() != null &&
-                                    !incomingTeam.getImageUrl().equals(existingTeam.getImageUrl()));
-
-                    if (!teamChanged) {
-                        return existingTeam;
-                    }
-
-                    Team updatedTeam = Team.builder()
-                            .id(existingTeam.getId())
-                            .externalId(existingTeam.getExternalId())
-                            .name(incomingTeam.getName())
-                            .imageUrl(incomingTeam.getImageUrl())
-                            .build();
-
-                    Team savedTeam = teamRepositoryPort.save(updatedTeam);
-                    log.info(
-                            "Updated team in DB externalId={} name={} id={}",
-                            savedTeam.getExternalId(),
-                            savedTeam.getName(),
-                            savedTeam.getId()
-                    );
-                    return savedTeam;
-                })
-                .orElseGet(() -> {
-                    Team savedTeam = teamRepositoryPort.save(
-                            Team.builder()
-                                    .id(UUID.randomUUID())
-                                    .externalId(incomingTeam.getExternalId())
-                                    .name(incomingTeam.getName())
-                                    .imageUrl(incomingTeam.getImageUrl())
-                                    .build());
-
-                    log.info(
-                            "Created team in DB externalId={} name={} id={}",
-                            savedTeam.getExternalId(),
-                            savedTeam.getName(),
-                            savedTeam.getId()
-                    );
-                    return savedTeam;
-                });
-    }
-
     private Match saveOrUpdateMatch(Match externalMatch, Tournament tournament) {
-        Team savedHomeTeam = saveOrGetTeam(externalMatch.getHomeTeam());
-        Team savedAwayTeam = saveOrGetTeam(externalMatch.getAwayTeam());
+        Team savedHomeTeam = teamSyncService.saveOrGetTeam(externalMatch.getHomeTeam());
+        Team savedAwayTeam = teamSyncService.saveOrGetTeam(externalMatch.getAwayTeam());
 
         return matchRepositoryPort.findByExternalId(externalMatch.getExternalId())
                 .map(existingMatch -> {
@@ -132,10 +76,6 @@ public class MatchResultSyncService {
 
     private String buildMatchName(Team homeTeam, Team awayTeam) {
         return homeTeam.getName() + " vs " + awayTeam.getName();
-    }
-
-    private boolean isBlank(String value) {
-        return value == null || value.trim().isEmpty();
     }
 
     public void syncAllCompetitions() {
